services/queue-management-agent/pyproject.toml
[project]
name = "queue-management-agent"
version = "0.1.0"
description = "A queue management agent that can optimize crowd flow, reduce wait times, and provide staff allocation recommendations using camera feeds and historical data analysis."
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.8.0,<2.0.0",
    "google-cloud-aiplatform[adk,agent_engine]>=1.100.0",
    "google-cloud-bigquery>=3.0.0",
    "google-cloud-storage>=2.0.0",
    "python-dotenv>=1.0.0",
    "pathlib",
    "asyncio",
    "numpy>=1.20.0",
    "pandas>=1.3.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false 